requirements.txt
psutil
GPUtil
pystray
pillow
pyserial